# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Privet API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/memory/facts": {
      "post": {
        "tags": [
          "memory"
        ],
        "summary": "Create Fact",
        "description": "Store a new fact for a user.",
        "operationId": "create_fact_api_v1_memory_facts_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FactCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserFact"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/memory/facts/search": {
      "post": {
        "tags": [
          "memory"
        ],
        "summary": "Search Facts",
        "description": "Search for relevant facts using semantic similarity.",
        "operationId": "search_facts_api_v1_memory_facts_search_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FactSearch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/FactSearchResult"
                  },
                  "type": "array",
                  "title": "Response Search Facts Api V1 Memory Facts Search Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/memory/facts/{user_id}": {
      "get": {
        "tags": [
          "memory"
        ],
        "summary": "Get User Facts",
        "description": "Get all facts for a user.",
        "operationId": "get_user_facts_api_v1_memory_facts__user_id__get",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserFact"
                  },
                  "title": "Response Get User Facts Api V1 Memory Facts  User Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/memory/facts/{fact_id}": {
      "delete": {
        "tags": [
          "memory"
        ],
        "summary": "Delete Fact",
        "description": "Delete a specific fact.",
        "operationId": "delete_fact_api_v1_memory_facts__fact_id__delete",
        "parameters": [
          {
            "name": "fact_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Fact Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Delete Fact Api V1 Memory Facts  Fact Id  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/memory/summaries/generate": {
      "post": {
        "tags": [
          "memory"
        ],
        "summary": "Generate Profile Summary",
        "description": "Generate or update user profile summary.",
        "operationId": "generate_profile_summary_api_v1_memory_summaries_generate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileSummaryCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/memory/summaries/{user_id}": {
      "get": {
        "tags": [
          "memory"
        ],
        "summary": "Get Profile Summaries",
        "description": "Get all profile summaries for a user.",
        "operationId": "get_profile_summaries_api_v1_memory_summaries__user_id__get",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProfileSummary"
                  },
                  "title": "Response Get Profile Summaries Api V1 Memory Summaries  User Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/memory/context/{user_id}": {
      "get": {
        "tags": [
          "memory"
        ],
        "summary": "Get Memory Context",
        "description": "Get complete memory context for a user.",
        "operationId": "get_memory_context_api_v1_memory_context__user_id__get",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          },
          {
            "name": "query",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Query"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemoryContext"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/memory/stats/{user_id}": {
      "get": {
        "tags": [
          "memory"
        ],
        "summary": "Get Memory Stats",
        "description": "Get memory statistics for a user.",
        "operationId": "get_memory_stats_api_v1_memory_stats__user_id__get",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemoryStats"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FactCreate": {
        "properties": {
          "user_id": {
            "type": "integer",
            "title": "User Id",
            "description": "User ID"
          },
          "fact_text": {
            "type": "string",
            "title": "Fact Text",
            "description": "Fact text"
          },
          "fact_summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fact Summary",
            "description": "Fact summary"
          },
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category",
            "description": "Fact category"
          },
          "source_message_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Message Id",
            "description": "Source message ID"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "fact_text"
        ],
        "title": "FactCreate",
        "description": "Schema for creating a fact."
      },
      "FactSearch": {
        "properties": {
          "user_id": {
            "type": "integer",
            "title": "User Id",
            "description": "User ID"
          },
          "query": {
            "type": "string",
            "title": "Query",
            "description": "Search query"
          },
          "limit": {
            "type": "integer",
            "maximum": 50.0,
            "minimum": 1.0,
            "title": "Limit",
            "description": "Maximum results",
            "default": 5
          },
          "min_similarity": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Min Similarity",
            "description": "Minimum similarity score",
            "default": 0.7
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "query"
        ],
        "title": "FactSearch",
        "description": "Schema for searching facts."
      },
      "FactSearchResult": {
        "properties": {
          "fact": {
            "$ref": "#/components/schemas/UserFact",
            "description": "The fact"
          },
          "similarity": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Similarity",
            "description": "Similarity score"
          },
          "relevance_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Relevance Reason",
            "description": "Why this fact is relevant"
          }
        },
        "type": "object",
        "required": [
          "fact",
          "similarity"
        ],
        "title": "FactSearchResult",
        "description": "Fact search result."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "MemoryContext": {
        "properties": {
          "user_id": {
            "type": "integer",
            "title": "User Id",
            "description": "User ID"
          },
          "relevant_facts": {
            "items": {
              "$ref": "#/components/schemas/FactSearchResult"
            },
            "type": "array",
            "title": "Relevant Facts",
            "description": "Relevant facts"
          },
          "profile_summaries": {
            "items": {
              "$ref": "#/components/schemas/ProfileSummary"
            },
            "type": "array",
            "title": "Profile Summaries",
            "description": "Profile summaries"
          },
          "recent_messages": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Recent Messages",
            "description": "Recent conversation messages"
          },
          "context_timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Context Timestamp",
            "description": "Context generation timestamp"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "relevant_facts",
          "profile_summaries",
          "recent_messages",
          "context_timestamp"
        ],
        "title": "MemoryContext",
        "description": "Complete memory context for a user."
      },
      "MemoryStats": {
        "properties": {
          "user_id": {
            "type": "integer",
            "title": "User Id",
            "description": "User ID"
          },
          "total_facts": {
            "type": "integer",
            "title": "Total Facts",
            "description": "Total facts stored"
          },
          "fact_categories": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Fact Categories",
            "description": "Facts by category"
          },
          "profile_summaries_count": {
            "type": "integer",
            "title": "Profile Summaries Count",
            "description": "Number of profile summaries"
          },
          "memory_density": {
            "type": "number",
            "minimum": 0.0,
            "title": "Memory Density",
            "description": "Memory density score"
          },
          "last_fact_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Fact Date",
            "description": "Date of last fact"
          },
          "storage_size_bytes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Storage Size Bytes",
            "description": "Storage size used"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "total_facts",
          "fact_categories",
          "profile_summaries_count",
          "memory_density"
        ],
        "title": "MemoryStats",
        "description": "Memory statistics for a user."
      },
...
      "ProfileSummary": {
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "Creation timestamp"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At",
            "description": "Last update timestamp"
          },
          "id": {
            "type": "integer",
            "title": "Id",
            "description": "Summary ID"
          },
          "user_id": {
            "type": "integer",
            "title": "User Id",
            "description": "User ID"
          },
          "summary_text": {
            "type": "string",
            "title": "Summary Text",
            "description": "Summary text"
          },
          "summary_topic": {
            "type": "string",
            "title": "Summary Topic",
            "description": "Summary topic"
          },
          "last_updated": {
            "type": "string",
            "format": "date-time",
            "title": "Last Updated",
            "description": "Last update timestamp"
          }
        },
        "type": "object",
        "required": [
          "created_at",
          "id",
          "user_id",
          "summary_text",
          "summary_topic",
          "last_updated"
        ],
        "title": "ProfileSummary",
        "description": "User profile summary schema."
      },
      "ProfileSummaryCreate": {
        "properties": {
          "user_id": {
            "type": "integer",
            "title": "User Id",
            "description": "User ID"
          },
          "summary_topic": {
            "type": "string",
            "title": "Summary Topic",
            "description": "Summary topic",
            "default": "general"
          },
          "force_regenerate": {
            "type": "boolean",
            "title": "Force Regenerate",
            "description": "Force regeneration",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "user_id"
        ],
        "title": "ProfileSummaryCreate",
        "description": "Schema for creating/updating profile summary."
      },
...
      "UserFact": {
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "Creation timestamp"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At",
            "description": "Last update timestamp"
          },
          "id": {
            "type": "integer",
            "title": "Id",
            "description": "Fact ID"
          },
          "user_id": {
            "type": "integer",
            "title": "User Id",
            "description": "User ID"
          },
          "fact_text": {
            "type": "string",
            "title": "Fact Text",
            "description": "Fact text"
          },
          "fact_summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fact Summary",
            "description": "Fact summary"
          },
          "source_message_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Message Id",
            "description": "Source message ID"
          },
          "confidence": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence",
            "description": "Fact confidence",
            "default": 1.0
          },
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category",
            "description": "Fact category"
          }
        },
        "type": "object",
        "required": [
          "created_at",
          "id",
          "user_id",
          "fact_text"
        ],
        "title": "UserFact",
        "description": "User fact schema."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}